Split the TOC state from the components to avoid cyclic dependencies on Tabs component.
