+++ apps/docs/features/ui/Tabs.tsx
@@ -3,7 +3,7 @@
 import { useCallback, type ComponentPropsWithoutRef, type PropsWithChildren } from 'react'
 import { Tabs as TabsPrimitive, type TabsProps } from 'ui'
 import { withQueryParams, withSticky, type QueryParamsProps } from 'ui-patterns/ComplexTabs'
-import { useTocRerenderTrigger } from '~/features/docs/GuidesMdx.client'
+import { useTocRerenderTrigger } from '~/features/docs/GuidesMdx.state'
 
 const TabsWithStickyAndQueryParams = withSticky<PropsWithChildren<TabsProps & QueryParamsProps>>(
   withQueryParams(TabsPrimitive)
@@ -40,4 +40,4 @@ const Tabs = ({
   )
 }
 
-export { Tabs, TabPanel }
+export { TabPanel, Tabs }
